+++ src/plonk/error.rs
@@ -1,6 +1,8 @@
+use std::io;
+
 /// This is an error that could occur during proving or circuit synthesis.
 // TODO: these errors need to be cleaned up
-#[derive(Debug, PartialEq)]
+#[derive(Debug)]
 pub enum Error {
     /// This is an error that can occur during synthesis of the circuit, for
     /// example, when the witness is not present.
@@ -14,7 +16,7 @@ pub enum Error {
     /// Opening error
     OpeningError,
     /// Transcript error
-    TranscriptError,
+    TranscriptError(io::Error),
     /// Instance provided has more rows than supported by circuit
     NotEnoughRowsAvailable,
     /// Instance provided exceeds number of available rows
@@ -22,4 +24,11 @@ pub enum Error {
     /// Circuit synthesis requires global constants, but circuit configuration did not
     /// call [`ConstraintSystem::enable_constant`] on fixed columns with sufficient space.
     NotEnoughColumnsForConstants,
+}
+
+impl From<io::Error> for Error {
+    fn from(error: io::Error) -> Self {
+        // The only place we can get io::Error from is the transcript.
+        Error::TranscriptError(error)
+    }
 }